ms_catalogo/requirements.txt
requests 
flask
sqlalchemy
psycopgflask-sqlalchemy
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
